requirements.txt
torch==1.12.0
onnx==1.12.0
numpy
opencv-python
rknn-toolkit2==1.4.0
